requirements.txt
fastapi
passlib[bcrypt]
pydantic
python-jose[cryptography]
python-multipart
redis[hiredis]
sqlalchemy
starlette
uvicorn
